Refactored Directory structure to match maven/gradle builds
