+++ build.gradle
@@ -1,0 +1,9 @@
+apply plugin: 'java'
+
+dependencies {
+    implementation 'org.junit.jupiter:junit-jupiter:5.9.0'
+}
+
+repositories {
+    mavenCentral()
+}
