+++ src/main/java/com/boyd/gofish/InteractivePlayer.java
@@ -1,12 +1,13 @@
-package GoFish;
+package com.boyd.gofish;
+
+import com.boyd.deckofcards.Card;
+import com.boyd.deckofcards.Card.Rank;
+import com.boyd.deckofcards.DeckOfCards;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Optional;
 import java.util.Scanner;
-import com.boyd.deckofcards.Card;
-import com.boyd.deckofcards.DeckOfCards;
-import com.boyd.deckofcards.Card.Rank;
 
 public class InteractivePlayer extends Player implements PlayerInterface {
