+++ src/test/GoFishTest.java
@@ -1,15 +1,17 @@
 package test;
 
-import java.util.ArrayList;
-import java.util.HashMap;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-import com.boyd.deckofcards.Card;
-import com.boyd.deckofcards.Card.Rank;
 import GoFish.GoFish;
 import GoFish.Player;
+import com.boyd.deckofcards.Card;
+import com.boyd.deckofcards.Card.Rank;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+import java.util.ArrayList;
+import java.util.HashMap;
+
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 public class GoFishTest {
